Refactor UserAuth components
