--- src/components/UserAuth/SubComponents/EmailLogIn/EmailLogIn.module.scss
@@ -1,38 +1,0 @@
-@use "../../../../variables.scss" as v;
-
-// General form styles
-.form {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  color: v.$clr-font-light;
-}
-
-// Form group styles
-.formGroup {
-  margin-bottom: 1em;
-  width: 100%;
-}
-
-// Text label styles
-.textLabel {
-  display: block;
-  font-size: var(--fs-extra-big);
-  font-weight: v.$fw-bold;
-  margin-bottom: 0.3em;
-}
-
-// Text input styles
-.textInput {
-  width: 100%;
-  padding: 0.5em 1em;
-  border: 1px solid v.$clr-font-light;
-  border-radius: 5px;
-  background-color: inherit;
-  font-size: var(--fs-big);
-
-  // Placeholder styles
-  &::placeholder {
-    font-size: var(--fs-small);
-  }
-}
