--- src/components/UserAuth/SubComponents/EmailSignUp/EmailSignUp.module.scss
@@ -1,9 +1,0 @@
-@use "../../../../../src/variables.scss" as v;
-
-// General form styles
-.form {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  color: v.$clr-font-light;
-}
