--- src/components/UserAuth/SubComponents/InputGroup/InputGroup.module.scss
@@ -1,36 +1,0 @@
-@use "../../../../variables.scss" as v;
-
-// Form group styles
-.formGroup {
-  margin-bottom: 1em;
-  width: 100%;
-  max-width: 500px;
-}
-
-// Text label styles
-.textLabel {
-  display: block;
-  font-size: var(--fs-extra-big);
-  font-weight: v.$fw-bold;
-  margin-bottom: 0.3em;
-}
-
-// Text input styles
-.textInput {
-  width: 100%;
-  padding: 0.5em 1em;
-  border: 1px solid v.$clr-font-light;
-  border-radius: 5px;
-  background-color: inherit;
-  font-size: var(--fs-big);
-
-  // Placeholder styles
-  &::placeholder {
-    font-size: var(--fs-small);
-  }
-}
-
-.error {
-  color: v.$clr-accent;
-  font-size: var(--fs-reg);
-}
